+++ activity/brightpay/react/changelog.md
@@ -2651,3 +2651,14 @@
 
 ---
 
+# Brightpay Development - 2025-03-22
+
+**Commit #226**
+
+- Repository: brightpay/realhealth
+- Type: healthcare-platform
+- Language: react
+- Month: 3/2025
+
+---
+